move account-detail directive to shared/directives folder, update relevant imports. Move your-details to create-account folder
