+++ src/app/features/workplace/create-user-account/create-user-account.component.ts
@@ -11,7 +11,7 @@ import { BreadcrumbService } from '@core/services/breadcrumb.service';
 import { CreateAccountService } from '@core/services/create-account/create-account.service';
 import { ErrorSummaryService } from '@core/services/error-summary.service';
 import { EstablishmentService } from '@core/services/establishment.service';
-import { AccountDetailsDirective } from '@features/account/account-details/account-details';
+import { AccountDetailsDirective } from '@shared/directives/user/account-details.directive';
 
 @Component({
   selector: 'app-create-account',
