+++ src/app/features/workplace/user-account-edit-details/user-account-edit-details.component.ts
@@ -8,7 +8,7 @@ import { BackService } from '@core/services/back.service';
 import { BreadcrumbService } from '@core/services/breadcrumb.service';
 import { ErrorSummaryService } from '@core/services/error-summary.service';
 import { UserService } from '@core/services/user.service';
-import { AccountDetailsDirective } from '@features/account/account-details/account-details';
+import { AccountDetailsDirective } from '@shared/directives/user/account-details.directive';
 
 @Component({
   selector: 'app-user-account-edit-details',
